-Basic Recommit

+++ src/Core/GameResourceLoader.java
@@ -37,24 +37,24 @@ public class GameResourceLoader {
 			// Load tiles
 			tileMap = ImageIO.read(this.getClass().getClassLoader().getResourceAsStream("Tiles.png")); // Grab the tilemap
 
-			for (byte i = 0; i < tiles.length; i++) { // For every tile that there should be
+			for (int i = 0; i < tiles.length; i++) { // For every tile that there should be
 				tiles[i] = tileMap.getSubimage(i * 32, 0, 32, 32); // Create a subimage from tile map, and store it as a separate image
 			}
 
 			// Load plants
 			plantMap = ImageIO.read(this.getClass().getClassLoader().getResourceAsStream("Plant.png")); // Grab the tilemap
 
-			for (byte i = 0; i < plants.length; i++) { // For every tile that there should be
+			for (int i = 0; i < plants.length; i++) { // For every tile that there should be
 				plants[i] = plantMap.getSubimage(i * 32, 0, 32, 32); // Create a subimage from tile map, and store it as a seperate image
 			}
 
 			playerMap = ImageIO.read(this.getClass().getClassLoader().getResourceAsStream("Player.png"));
-			for (byte i = 0; i < playerFaces.length; i++) { // For every tile that there should be
+			for (int i = 0; i < playerFaces.length; i++) { // For every tile that there should be
 				playerFaces[i] = playerMap.getSubimage(i * 32, 0, 32, 32); // Create a subimage from tile map, and store it as a separate image
 			}
 
 			toolMap = ImageIO.read(this.getClass().getClassLoader().getResourceAsStream("Tools.png"));
-			for (byte i = 0; i < tools.length; i++) { // For every tile that there should be
+			for (int i = 0; i < tools.length; i++) { // For every tile that there should be
 				tools[i] = toolMap.getSubimage(i * 32, 0, 32, 32); // Create a subimage from tile map, and store it as a separate image
 			}
